See previous commit message.

+++ shared/duplicacyln.sh
@@ -81,7 +81,7 @@ sighandler() {
 	unwrap
 
 	CLI="$(newest "$WEBUI_BIN")"
-	if [ -x "$WEBUI_BIN/$CLI" ]; then
+	if [ -x "$WEBUI_BIN/$CLI" -a ! -h "$WEBUI_BIN/$CLI" ]; then
 		rm -f /usr/bin/duplicacy
 		ln -s "$WEBUI_BIN/$CLI" /usr/bin/duplicacy
 		log_alert "[DuplicacyLog] /usr/bin/duplicacy linked to $WEBUI_BIN/$CLI" 0
